fastapi_events/requirements.txt
fastapi
uvicorn[standard]
pydantic
kafka-python
redis
celery 
pyjwt
python-dotenv
boto3